backend/requirements.txt
flask==2.3.3
flask-sqlalchemy==3.0.5
flask-cors==4.0.0
pymysql==1.1.0
cryptography==41.0.4
redis==5.0.1
pyjwt==2.8.0
python-dotenv==1.0.0
werkzeug==2.3.7